--- src/components/DescriptionForm/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-import "./styles.css";
-
-export default function DescriptionForm({ description }) {
-    return (
-        <p className="form__description">
-            {description}
-        </p>
-    )
-};
