+++ src/components/DescriptionForm/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './styles.css';
+
+export default function DescriptionForm({ description }) {
+  return (
+    <p className="form__description">
+      {description}
+    </p>
+  );
+}
