--- src/components/HeaderForm/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-import "./styles.css";
-
-export default function HeaderForm({ title }) {
-    return (
-        <p className="form__header">
-            {title}
-        </p>
-    )
-};
