+++ src/components/HeaderForm/index.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import './styles.css';
+
+export default function HeaderForm({ title }) {
+  return (
+    <p className="form__header">
+      {title}
+    </p>
+  );
+}
