--- src/components/InputOutlineForm/index.js
@@ -1,20 +1,0 @@
-import React from "react";
-
-import "./styles.css";
-
-export default function InputOutlineForm({
-    inputPlaceholder,
-    onChangeInput,
-    inputValue,
-    inputType = "text",
-}) {
-    return (
-        <input
-            className="form__input"
-            type={inputType}
-            placeholder={inputPlaceholder}
-            value={inputValue}
-            onChange={event => onChangeInput(event.target.value)}
-        />
-    )
-};
